+++ disassemblers/ofrak_angr/src/ofrak_angr/components/blocks/unpackers.py
@@ -17,8 +17,7 @@ from ofrak.core.data import DataWord
 from ofrak.resource import Resource
 from ofrak.service.resource_service_i import ResourceFilter
 from ofrak_angr.components.angr_analyzer import AngrAnalyzerConfig, AngrCodeRegionModifier
-from ofrak_angr.components.identifiers import AngrAnalysisResource
-from ofrak_angr.model import AngrAnalysis
+from ofrak_angr.model import AngrAnalysis, AngrAnalysisResource
 
 LOGGER = logging.getLogger(__name__)