+++ disassemblers/ofrak_binary_ninja/CHANGELOG.md
@@ -6,7 +6,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/) 
 ## [Unreleased 0.1.1](https://github.com/redballoonsecurity/ofrak/tree/master)
 
 ### Added
-- Support `ProgramAttributes` `entry_points` and `base_address` fields for passing program metadata to Binary Ninja
+- Add `BinaryNinjaAutoLoadProject` / `BinaryNinjaCustomLoadProject` tags and `BinaryNinjaCustomLoadAnalyzer` for custom binary loading with `ProgramAttributes` metadata
 
 ## 0.1.0 - 2022-01-25
 ### Added